+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Fix cho-jung bug `웬 + ㅊ$ㅜ = 웬ㅊ`
 * Fix composable jungseong bug `ㅇ + $ㅆ$ㅜ + $ㅊ$ㅔ +  = 웇`
+* Fix xim bug [#304](https://github.com/Riey/kime/issues/304)
 
 ## 1.1.2