+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Improve
 
 * Fix kime-check fail [#307](https://github.com/Riey/kime/issues/307)
+* Fix preedit string disapear when press hotkey [#310](https://github.com/Riey/kime/issues/310)
 
 ## 1.1.3