+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
 * Make character typing order strict `ㅏ + ㄱ = ㅏㄱ`
 * Add `TreatJongseongAsChoseong` addon
 * Add `FlexibleComposeOrder` addon [#318](https://github.com/Riey/kime/issues/318)
+* Check LANG env in kime-check [#317](https://github.com/Riey/kime/issues/317)
 
 ## 1.1.3