+++ docs/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Improve
 
+## 1.2.0
+
+### Improve
+
 * Fix kime-check fail [#307](https://github.com/Riey/kime/issues/307)
 * Fix preedit string disapear when press hotkey [#310](https://github.com/Riey/kime/issues/310)
 * Make character typing order strict `ㅏ + ㄱ = ㅏㄱ`